+++ src/impls/matmul_cuda.h
@@ -38,9 +38,9 @@ struct matmulScalesCUDA
 };
 
 template<typename InputType, typename OutputType, int version>
-void matmulTensorCores(InputType *A, InputType *B, OutputType *C, int M, int K, int N);
+void matmulTensorCores(InputType *A, InputType *B, OutputType *C, size_t M, size_t K, size_t N);
 
 template<typename InputType, typename OutputType, int version>
-void matmulCUDACores(InputType *A, InputType *B, OutputType *C, int M, int K, int N);
+void matmulCUDACores(InputType *A, InputType *B, OutputType *C, size_t M, size_t K, size_t N);
 
 #endif // MATMUL_CUDA_H
