+++ src/impls/matmul_reference.cu
@@ -1,7 +1,7 @@
 #include "../timer.h"
 #include "../matmul.h"
 
-flop_counts matmul_reference32(float *A, float *B, float *C, int M, int K, int N) {
+flop_counts matmul_reference32(float *A, float *B, float *C, size_t M, size_t K, size_t N) {
     for (int i = 0; i < M; i++) {
         for (int j = 0; j < N; j++) {
             float c_ij = 0.0;
@@ -15,7 +15,7 @@ flop_counts matmul_reference32(float *A, float *B, float *C, int M, int K, int N
     return {0L, 2L*M*K*N, 0L};
 }
 
-flop_counts matmul_reference64(double *A, double *B, double *C, int M, int K, int N) {
+flop_counts matmul_reference64(double *A, double *B, double *C, size_t M, size_t K, size_t N) {
     for (int i = 0; i < M; i++) {
         for (int j = 0; j < N; j++) {
             double c_ij = 0.0;
